refactor(admin): split simulation run execution responsibilities
