+++ docs/Workflows-and-Troubleshooting.md
@@ -50,7 +50,7 @@ The command-line interface remains **fully backward compatible** with previous v
 Run a simulation using a `.scenario` file:
 
 ```bash
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.scenario.ScenarioRunnerMain \
   path/to/scenario.scenario
 ```
@@ -64,12 +64,12 @@ If no scenario file is provided, uses `demo.scenario` from the classpath.
 **Example:**
 ```bash
 # Run with a specific scenario file
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.scenario.ScenarioRunnerMain \
   my-test-scenario.scenario
 
 # Run with default demo scenario
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.scenario.ScenarioRunnerMain
 ```
 
@@ -113,7 +113,7 @@ Event rows are comma-delimited, and each request carries a unique alias (`p1`, `
 Run a quick demo simulation with built-in configuration:
 
 ```bash
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.Main \
   --strategy=directional-scan
 ```
@@ -391,7 +391,7 @@ You can reproduce any UI-driven run using the CLI by downloading the generated i
 
 4. **Run via CLI**
    ```bash
-   java -cp target/lift-simulator-0.57.7.jar \
+   java -cp target/lift-simulator-0.57.8.jar \
      com.liftsimulator.scenario.ScenarioRunnerMain \
      run-42-reproduction.scenario
    ```
@@ -436,7 +436,7 @@ curl -X POST http://localhost:8080/api/batch/generate-input \
 **3. Run the scenario:**
 
 ```bash
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.scenario.ScenarioRunnerMain \
   run-42-reproduction.scenario
 ```
@@ -528,7 +528,7 @@ Where `run-123-inputs.json` contains:
 **2. Run via CLI:**
 
 ```bash
-java -cp target/lift-simulator-0.57.7.jar \
+java -cp target/lift-simulator-0.57.8.jar \
   com.liftsimulator.scenario.ScenarioRunnerMain \
   morning-rush-reproduction.scenario
 ```
