+++ frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "lift-simulator-admin",
   "private": true,
-  "version": "0.57.7",
+  "version": "0.57.8",
   "type": "module",
   "engines": {
     "node": "^20.19.0 || >=22.12.0"
