+++ pom.xml
@@ -14,7 +14,7 @@
 
     <groupId>com.liftsimulator</groupId>
     <artifactId>lift-simulator</artifactId>
-    <version>0.57.7</version>
+    <version>0.57.8</version>
     <packaging>jar</packaging>
 
     <name>Lift Simulator</name>
